# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  basket.BasketItemCommand:
    properties:
      basketItemId:
        type: integer
      productId:
        type: integer
      productVariantId:
        type: integer
      quantity:
        maximum: 1000
        type: integer
    required:
    - productId
    - quantity
    type: object
  basket.UpdateBasketCommand:
    properties:
      basketItems:
        items:
          $ref: '#/definitions/basket.BasketItemCommand'
        minItems: 1
        type: array
      code:
        type: string
      simpleAdd:
        type: boolean
      siteId:
        type: integer
    required:
    - basketItems
    - siteId
    type: object
...
  utils.Result:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
      success:
        type: boolean
      systemMessage:
        type: string
    type: object
...
host: localhost:8585
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is a sample server.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: My API
  version: "1.0"
paths:
...
  /basket:
    get:
      consumes:
      - application/json
      description: Retrieves the current user's shopping basket
      parameters:
      - in: query
        name: siteId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Basket details
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "404":
          description: Basket not found
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Get current basket
      tags:
      - basket
    put:
      consumes:
      - application/json
      description: Updates the user's shopping basket with new items or quantities
      parameters:
      - description: Updated basket information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/basket.UpdateBasketCommand'
      produces:
      - application/json
      responses:
        "200":
          description: Updated basket
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Update shopping basket
      tags:
      - basket
  /basket/admin/all:
    get:
      consumes:
      - application/json
      description: Admin endpoint to retrieve all shopping baskets across all users
      parameters:
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - in: query
        name: search
        type: string
      - enum:
        - title
        - description
        in: query
        name: searchBy
        type: string
      - enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - enum:
        - title
        - description
        in: query
        name: sortBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of all user baskets
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: 'Admin: Get all user baskets'
      tags:
      - basket
  /basket/user/all:
    get:
      consumes:
      - application/json
      description: Retrieves all shopping baskets for the current user
      parameters:
      - in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: pageSize
        required: true
        type: integer
      - in: query
        name: search
        type: string
      - enum:
        - title
        - description
        in: query
        name: searchBy
        type: string
      - in: query
        name: siteId
        required: true
        type: integer
      - enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - enum:
        - title
        - description
        in: query
        name: sortBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of user baskets
          schema:
            $ref: '#/definitions/utils.Result'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/utils.Result'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.Result'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Result'
      security:
      - BearerAuth: []
      summary: Get all user baskets
      tags:
      - basket
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"